Not continuing this project because I shouldn't be wasting time redoing the wheel... Although I'd love do finish this when I have free time

+++ src/Token.cpp
@@ -2,4 +2,4 @@
 
 sci::Token::Token() {}
 sci::Token::Token(TokenType type) : type(type) {}
-sci::Token::Token(TokenType type, const char* value, size_t valueSize) : type(type), value(value), valueSize(valueSize) {}
+sci::Token::Token(TokenType type, const char* value, uint64_t valueSize) : type(type), value(value), valueSize(valueSize) {}
